challenge-1/requirements.txt
pandas>=1.3.0
torch>=1.12.0
torchvision>=0.13.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.60.0
Pillow>=8.2.0
kaggle>=1.5.12
